+++ src/ModalPreviewLayout/test/ModalPreviewLayout.tsx
@@ -7,7 +7,7 @@ import * as enzyme from 'enzyme';
 import * as puppeteer from 'puppeteer';
 
 function ModalPreviewLayoutWithMandatoryProps() {
-  return <ModalPreviewLayout onClose={() => {}}>asd</ModalPreviewLayout>;
+  return <ModalPreviewLayout onClose={() => {}} children="text" />;
 }
 
 function ModalPreviewLayoutWithAllProps() {
@@ -17,27 +17,30 @@ function ModalPreviewLayoutWithAllProps() {
       actions={<div />}
       dataHook="hook"
       title="title"
-      shouldCloseOnOverlayClick>
-      asd
-    </ModalPreviewLayout>
+      shouldCloseOnOverlayClick
+      closeButtonTooltipText="text"
+      nextButtonTooltipText="text"
+      prevButtonTooltipText="text"
+      children="text"
+    />
   );
 }
 
 async function testkits() {
   const testkit = modalPreviewLayoutTestkitFactory({
     dataHook: 'hook',
-    wrapper: document.createElement('div')
+    wrapper: document.createElement('div'),
   });
 
   const enzymeTestkit = modalPreviewLayoutEnzymeTestkitFactory({
     dataHook: 'hook',
-    wrapper: enzyme.mount(<div />)
+    wrapper: enzyme.mount(<div />),
   });
 
   const browser = await puppeteer.launch();
   const page = await browser.newPage();
   const puppeteerTestkit = await modalPreviewLayoutPuppeteerTestkitFactory({
     dataHook: 'hook',
-    page
+    page,
   });
 }