+++ CHANGELOG.md
@@ -18,6 +18,8 @@ Types of changes:
 
 ### Added
 - `<RadioGroup/>` - Added UniDrivers to `RadioGroup` and `RadioButton`, fixed a bug in the existing `RadioButton` driver that prevented it from returing correct values. [#5459](https://github.com/wix/wix-style-react/pull/5459)
+- `<ModalPreviewLayout/>` - Fix props types [#5482](https://github.com/wix/wix-style-react/pull/5482)
+
 
 ## 8.11.0 - 2020-05-19
 ### Added